Sample and Sampler

+++ src/film.cpp
@@ -5,19 +5,19 @@
 
 #include "imageExportor.hpp"
 
-Film::Film(int width, int height) :
+Film::Film(size_t width, size_t height) :
     width_{width},
     height_{height},
-    buffer_{static_cast<std::size_t>(width) * height} {
+    buffer_{width * height} {
 }
 
-void Film::setResolution(int width, int height) noexcept {
+void Film::setResolution(size_t width, size_t height) noexcept {
     width_ = width;
     height_ = height;
     buffer_.resize(width_ * height_);
 }
 
-void Film::commit(int x, int y, const ColorRGB &color) {
+void Film::commit(size_t x, size_t y, const ColorRGB &color) {
     if (x < 0 || x >= width_ || y < 0 || y >= height_) {
         std::stringstream ss;
         ss << "Cannot commit color to pixel ("
